// workspace/src/lib/data/types.ts
export type StatusIbuHamil = 'Normal' | 'Risiko KEK';

export interface IbuHamil {
  id: number;
  nik: string;
  namaLengkap: string;
  usiaKehamilanMinggu: number;
  htp: string; // 'YYYY-MM-DD'
  lilaTerakCm: number;
  beratBadanKg?: number;
  tekananDarah?: string;
  status: StatusIbuHamil;
}